requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antiorm==1.2.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
cohere==4.11.2
colorama==0.4.6
datasets==2.13.1
db==0.1.1
db-sqlite3==0.0.1
dill==0.3.6
dnspython==2.3.0
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.16.3
idna==3.4
importlib-metadata==6.7.0
loguru==0.7.0
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.4
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pinecone-client==2.2.2
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
win32-setctime==1.1.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
